Remove custom avatar
